feat: refactor import, change name of components, refactor react examples to show use case
