+++ data/js/algorithms.js
