+++ data/react/basics.js
@@ -2,9 +2,9 @@ export const reactBasicsExamples = [
   {
     id: "jsx",
     name: "JSX Syntax",
-    description: "JSX allows you to write HTML-like syntax inside JavaScript.",
-    code: `// JSX Example
-  const element = <h1>Hello, React!</h1>;`,
+    description:
+      "JSX allows you to write HTML-like code within JavaScript. It gets compiled to React.createElement calls under the hood.",
+    code: `const element = <h1>Hello, React!</h1>;`,
     usage: `const element = <h1>Hello, React!</h1>;
   const result = element.type; // "h1"`,
     run: () => {
@@ -15,12 +15,12 @@ export const reactBasicsExamples = [
   {
     id: "component",
     name: "Function Component",
-    description: "React components are functions that return JSX.",
-    code: `// Functional Component
-  function Welcome() {
+    description:
+      "React components are JavaScript functions that return JSX. They're the building blocks of any React UI.",
+    code: `function Welcome() {
     return <p>Welcome to React!</p>;
   }`,
-    usage: `function Welcome() {
+    usage: `  function Welcome() {
     return <p>Welcome to React!</p>;
   }
   const element = Welcome();
@@ -35,11 +35,12 @@ export const reactBasicsExamples = [
   {
     id: "props",
     name: "Props",
-    description: "Props are inputs passed to components.",
+    description:
+      "Props are arguments passed into components, allowing data to flow from parent to child components.",
     code: `function Greet(props) {
     return <h2>Hello, {props.name}!</h2>;
   }`,
-    usage: `const Greet = ({ name }) => <h2>Hello, {name}!</h2>;
+    usage: `  const Greet = ({ name }) => <h2>Hello, {name}!</h2>;
   const element = Greet({ name: "Alice" });
   const result = element.props.children.join(" "); // "Hello, Alice!"`,
     run: () => {
@@ -51,7 +52,8 @@ export const reactBasicsExamples = [
   {
     id: "useState",
     name: "useState",
-    description: "useState lets you add local state to functional components.",
+    description:
+      "`useState` is a React Hook that lets you add and manage local state in function components.",
     code: `import { useState } from "react";
   
   function Counter() {
@@ -62,7 +64,7 @@ export const reactBasicsExamples = [
       </button>
     );
   }`,
-    usage: `let count = 0;
+    usage: `  let count = 0;
   const setCount = (val) => { count = val; };
   setCount(count + 1);
   const result = count; // 1`,
@@ -77,13 +79,13 @@ export const reactBasicsExamples = [
     id: "useEffect",
     name: "useEffect",
     description:
-      "useEffect runs side effects after rendering (e.g., data fetching).",
+      "`useEffect` is a React Hook used for side effects like data fetching, subscriptions, or manually changing the DOM.",
     code: `import { useEffect } from "react";
   
   useEffect(() => {
     console.log("Component mounted");
   }, []);`,
-    usage: `let message = "";
+    usage: `  let message = "";
   const mockEffect = () => {
     message = "Component mounted";
   };
